requirements.txt
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
passlib
marshmallow
python-dotenv
flask-migrate
gunicorn
psycopg2
requests